+++ packages/ocean-react-formik/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.130.0](https://github.com/ocean-ds/ocean-web/compare/v1.129.0...v1.130.0) (2026-02-12)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.129.0](https://github.com/ocean-ds/ocean-web/compare/v1.128.1...v1.129.0) (2026-02-05)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.128.0](https://github.com/ocean-ds/ocean-web/compare/v1.127.0...v1.128.0) (2026-02-04)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.127.0](https://github.com/ocean-ds/ocean-web/compare/v1.126.0...v1.127.0) (2026-01-14)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.126.0](https://github.com/ocean-ds/ocean-web/compare/v1.125.1...v1.126.0) (2026-01-14)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+## [1.125.1](https://github.com/ocean-ds/ocean-web/compare/v1.125.0...v1.125.1) (2026-01-13)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
+# [1.125.0](https://github.com/ocean-ds/ocean-web/compare/v1.124.0...v1.125.0) (2026-01-13)
+
+**Note:** Version bump only for package @useblu/ocean-react-formik
+
 # [1.124.0](https://github.com/ocean-ds/ocean-web/compare/v1.123.0...v1.124.0) (2026-01-08)
 
 **Note:** Version bump only for package @useblu/ocean-react-formik